+++ src/hictk/cli/cli_convert.cpp
@@ -116,7 +116,7 @@ void Cli::make_convert_subcommand() {
       c.compression_lvl,
       "Compression level used to compress interactions.\n"
       "Defaults to 6 and 10 for .cool and .hic files, respectively.")
-      ->check(CLI::Range(std::uint8_t{1}, MAX_HIC_COMPRESSION_LEVEL))
+      ->check(CLI::Range(std::int16_t{1}, MAX_HIC_COMPRESSION_LEVEL))
       ->capture_default_str();
   sc.add_flag(
       "--skip-all-vs-all,!--no-skip-all-vs-all",
